chore: rollback tmp dir renaming Vercel build issue
