+++ jobs/Frontend/fe-challenge/src/pages/loading/index.ts
