+++ jobs/Frontend/fe-challenge/src/pages/movie-detail/hooks/useGetMovieDetail.ts
