+++ jobs/Frontend/fe-challenge/src/pages/movie-detail/index.ts
