+++ jobs/Frontend/fe-challenge/src/pages/search/components/EmptySearch.tsx
