+++ jobs/Frontend/fe-challenge/src/pages/search/components/ResultsNumber.tsx
