+++ jobs/Frontend/fe-challenge/src/pages/search/hooks/useBuildLocationState.ts
