+++ jobs/Frontend/fe-challenge/src/pages/search/hooks/useQuerySearch.ts
