+++ jobs/Frontend/fe-challenge/src/pages/search/index.ts
